discord-daily-ingester/requirements.txt
functions-framework
requests
google-cloud-firestore
beautifulsoup4